feat: add time scale on date selector.
Now, the logic is following the next steps:
If the interval is daily or weekly, the setting is ```month``` for a single month display.

If the interval is monthly, the parameter is ```year``` for 12 months display.

If the interval is yearly, the parameter is ```multi-year``` to return the previous or subsequent years.