refactor: add two date selectors with their respective behavior.
One of the components corresponds to the old implementation of ```MatDateRangeSelectionStrategy``` that was associated with the timeline component, and the other component corresponds to the custom date selector.

To work, the ```PresetDateSelectorComponent```, has retrieved the refactored methods of ```TimelinePageComponent```, and recalls the ```TimelineService``` and ```DateRangeService``` classes.

For the ```CustomDateSelectorComponent```, it is a basic implementation of the DatePicker component with the necessary options.

The use of the ```Input()``` allows to give the date selectors the range and the list of dates, and the ```Output``` are in charge of returning this updated list to the timeline component for display.